requirements/base.txt
psycopg2-binary==2.9.3
pydantic==1.9.0
